feat: sweetalert2 in necessary places

+++ src/components/Elements.jsx
@@ -1,5 +1,6 @@
 import React from "react";
 import { useDispatch } from "react-redux";
+import Swal from "sweetalert2";
 import { deleteRegister } from "../actions/nomina";
 
 const Elements = ({ data }) => {
@@ -14,7 +15,27 @@ const Elements = ({ data }) => {
     fechaFormato = fecha;
   }
   const handleDelete = () => {
-    dispatch(deleteRegister(id));
+    Swal.fire({
+      title: "¿Estás seguro de eliminar este registro?",
+      text: "Si eliminas este registro no habrá forma de recuperarlo.",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      cancelButtonText: "Cancelar",
+      confirmButtonText: "Si",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        Swal.fire({
+          icon: "success",
+          title: "Eliminado",
+          text: "Sé ha eliminado correctamente.",
+          showConfirmButton: false,
+          timer: 1000,
+        });
+        dispatch(deleteRegister(id));
+      }
+    });
   };
   return (
     <>
@@ -22,7 +43,7 @@ const Elements = ({ data }) => {
       <td>${pago}</td>
       <td>
         <button onClick={handleDelete} className="btn red">
-        <i className="material-icons">delete_forever</i>
+          <i className="material-icons">delete_forever</i>
         </button>
       </td>
     </>
